EntryAndExitOnlyOnSecureConnection ( Particular Wifi MAC Address )

+++ app/src/main/java/com/example/xms/UserDashboardActivity.java
@@ -21,6 +21,7 @@ import androidx.core.content.ContextCompat;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
+import com.google.common.base.Strings;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.DocumentId;
@@ -32,8 +33,10 @@ import com.google.firebase.firestore.SetOptions;
 import com.google.firebase.firestore.auth.User;
 
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import static android.content.ContentValues.TAG;
@@ -43,11 +46,12 @@ public class UserDashboardActivity extends Activity {
     private Button profile,entry,exit,userLog,peoplePresent,social;
 
     private String fname,lname,pNum;
+    private String BSSID,probBSSID = "02:00:00:00:00:00";
 
     String timestamp;
     FirebaseFirestore fbd;
     FirebaseUser user;
-    DocumentReference persDetail;
+    DocumentReference persDetail,macAdd;
 
 
     public static final int requestCode = 2019123;
@@ -82,22 +86,46 @@ public class UserDashboardActivity extends Activity {
             }
         });
 
+        macAdd = fbd.collection("MAC-Addresses").document("MAC");
 
 
-//        ActivityCompat.shouldShowRequestPermissionRationale(UserDashboardActivity.this, Manifest.permission.ACCESS_WIFI_STATE);
+        List<String> details = new ArrayList<>();
+        macAdd.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+            @Override
+            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
+                if(task.isSuccessful()){
+                    DocumentSnapshot document = task.getResult();
+                    if(document.exists()){
+                        Map<String,Object> mp= document.getData();
+                        for (Map.Entry<String, Object> entry : mp.entrySet()) {
+                            details.add(entry.getKey().trim());
+                        }
+                    }
+                }
+            }
+        });
 
-        if( ContextCompat.checkSelfPermission(this,Manifest.permission.ACCESS_COARSE_LOCATION)!= PackageManager.PERMISSION_GRANTED) {
-            requestPermissions(new String[] {Manifest.permission.ACCESS_COARSE_LOCATION},requestCode);
-        }
-
-//        WifiManager wifiMgr = (WifiManager) getApplicationContext().getSystemService(Context.WIFI_SERVICE);                                  // TODO: Resolve the incosistent BSSID info
-//        WifiInfo wifiInfo = wifiMgr.getConnectionInfo();
-//        Toast.makeText(UserDashboardActivity.this, wifiInfo.toString()  ,Toast.LENGTH_LONG).show();
-
+        Log.i("Details",details.toString());
 
         entry.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+
+                if( ContextCompat.checkSelfPermission(UserDashboardActivity.this,Manifest.permission.ACCESS_FINE_LOCATION)!= PackageManager.PERMISSION_GRANTED) {
+                    requestPermissions(new String[] {Manifest.permission.ACCESS_FINE_LOCATION},requestCode);
+                }
+
+                if( ContextCompat.checkSelfPermission(UserDashboardActivity.this,Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
+                    WifiManager wifiMgr = (WifiManager) getApplicationContext().getSystemService(WIFI_SERVICE);
+                    WifiInfo wifiInfo = wifiMgr.getConnectionInfo();
+                    BSSID = wifiInfo.getBSSID().trim();
+                }else{
+                    Toast.makeText(UserDashboardActivity.this ,"Location Permission Needed",Toast.LENGTH_SHORT).show();
+                }
+
+                if(BSSID.equals(probBSSID) && ContextCompat.checkSelfPermission(UserDashboardActivity.this,Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED ) {
+                    Toast.makeText(UserDashboardActivity.this ,"Please Turn On Location",Toast.LENGTH_SHORT).show();
+                }
 
                 persDetail = fbd.collection("personPresent").document(user.getUid().trim());
                 persDetail.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
@@ -106,12 +134,16 @@ public class UserDashboardActivity extends Activity {
                         if(documentSnapshot.exists()){
                             Toast.makeText(UserDashboardActivity.this,"Already Inside",Toast.LENGTH_SHORT).show();
                         }else{
-                            timestamp = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss").format(new Date());
-                            Map<String,String> detail = new HashMap<>();
-                            detail.put(timestamp.trim(),"Entry");
-                            fbd.collection("UserLogDetail").document(user.getUid().trim()).set(detail, SetOptions.merge());
-                            fbd.collection("personPresent").document(user.getUid().trim()).set(detail);
-                            Toast.makeText(UserDashboardActivity.this,"Entry Detail Recorded",Toast.LENGTH_SHORT).show();
+                            if( !BSSID.equals(probBSSID) && ContextCompat.checkSelfPermission(UserDashboardActivity.this,Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED && details.contains(BSSID)==true) {
+                                timestamp = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss").format(new Date());
+                                Map<String, String> detail = new HashMap<>();
+                                detail.put(timestamp.trim(), "Entry");
+                                fbd.collection("UserLogDetail").document(user.getUid().trim()).set(detail, SetOptions.merge());
+                                fbd.collection("personPresent").document(user.getUid().trim()).set(detail);
+                                Toast.makeText(UserDashboardActivity.this, "Entry Detail Recorded", Toast.LENGTH_SHORT).show();
+                            }else{
+                                Toast.makeText(UserDashboardActivity.this,"Not on secure Connection to log the details",Toast.LENGTH_LONG).show();
+                            }
                         }
                     }
                 });
@@ -123,18 +155,37 @@ public class UserDashboardActivity extends Activity {
             @Override
             public void onClick(View v) {
 
+                if( ContextCompat.checkSelfPermission(UserDashboardActivity.this,Manifest.permission.ACCESS_FINE_LOCATION)!= PackageManager.PERMISSION_GRANTED) {
+                    requestPermissions(new String[] {Manifest.permission.ACCESS_FINE_LOCATION},requestCode);
+                }
+
+                if( ContextCompat.checkSelfPermission(UserDashboardActivity.this,Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
+                    WifiManager wifiMgr = (WifiManager) getApplicationContext().getSystemService(WIFI_SERVICE);
+                    WifiInfo wifiInfo = wifiMgr.getConnectionInfo();
+                    BSSID = wifiInfo.getBSSID().trim();
+                }else{
+                    Toast.makeText(UserDashboardActivity.this ,"Location Permission Needed",Toast.LENGTH_SHORT).show();
+                }
+
+                if(BSSID.equals(probBSSID) && ContextCompat.checkSelfPermission(UserDashboardActivity.this,Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED ) {
+                    Toast.makeText(UserDashboardActivity.this ,"Please Turn On Location",Toast.LENGTH_SHORT).show();
+                }
 
                 persDetail = fbd.collection("personPresent").document(user.getUid().trim());
                 persDetail.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
                     @Override
                     public void onSuccess(DocumentSnapshot documentSnapshot) {
                         if(documentSnapshot.exists()){
-                            timestamp = new SimpleDateFormat("dd-MM-yyyy kk:mm:ss").format(new Date());
-                            Map<String,String> detail = new HashMap<>();
-                            detail.put(timestamp.trim(),"Exit");
-                            fbd.collection("UserLogDetail").document(user.getUid().trim()).set(detail, SetOptions.merge());
-                            fbd.collection("personPresent").document(user.getUid().trim()).delete();
-                            Toast.makeText(UserDashboardActivity.this,"Exit Detail Recorded",Toast.LENGTH_SHORT).show();
+                            if( !BSSID.equals(probBSSID) && ContextCompat.checkSelfPermission(UserDashboardActivity.this,Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED && details.contains(BSSID)==true ) {
+                                timestamp = new SimpleDateFormat("dd-MM-yyyy kk:mm:ss").format(new Date());
+                                Map<String, String> detail = new HashMap<>();
+                                detail.put(timestamp.trim(), "Exit");
+                                fbd.collection("UserLogDetail").document(user.getUid().trim()).set(detail, SetOptions.merge());
+                                fbd.collection("personPresent").document(user.getUid().trim()).delete();
+                                Toast.makeText(UserDashboardActivity.this, "Exit Detail Recorded", Toast.LENGTH_SHORT).show();
+                            }else{
+                                Toast.makeText(UserDashboardActivity.this,"Not on secure Connection to log the details",Toast.LENGTH_LONG).show();
+                            }
                         }else{
                             Toast.makeText(UserDashboardActivity.this,"Already Outside",Toast.LENGTH_SHORT).show();
                         }
